ProjectSoftwareBackend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
Werkzeug
python-dotenv
